+++ src/message/message_type.rs
@@ -1,7 +1,0 @@
-//#[derive(PartialEq, Clone, Copy, Debug)]
-//pub enum MessageType {
-//    Other,
-//    ActorStopMessage,
-//    SystemStopMessage,
-//}
-//