+++ CHANGELOG.md
@@ -3,6 +3,8 @@
  - switch internal actor mailbox to `flume` from `crossbeam-channel`
    - `crossbeam-channel` is still being used to keep actors thread independent
    - see [github issue #13](https://github.com/sers-dev/tyra/issues/13) for reasoning
+ - added generic `ActorInitMessage` without a generic `Handler` implementation
+   - a `Handler<ActorInitMessage>` can optionally be implemented by the user for any `actor` which can then be used to initialize the actor
 
 # 0.7.0